ref: Move useUrlParams() to utls/url/ with simiar hooks
