+++ static/app/components/replays/virtualizedGrid/useDetailsSplit.tsx
@@ -1,8 +1,8 @@
 import type {RefObject} from 'react';
 import {useCallback} from 'react';
 
+import useUrlParams from 'sentry/utils/url/useUrlParams';
 import {useResizableDrawer} from 'sentry/utils/useResizableDrawer';
-import useUrlParams from 'sentry/utils/useUrlParams';
 
 interface OnClickProps {
   dataIndex: number;
