+++ static/app/utils/replays/hooks/useActiveReplayTab.tsx
@@ -1,6 +1,6 @@
 import {useCallback} from 'react';
 
-import useUrlParams from 'sentry/utils/useUrlParams';
+import useUrlParams from 'sentry/utils/url/useUrlParams';
 
 export enum TabKey {
   AI = 'ai',
