+++ static/app/views/replays/detail/errorList/useSortErrors.tsx
@@ -2,7 +2,7 @@ import {useCallback, useMemo} from 'react';
 
 import type {SortConfig} from 'sentry/components/replays/virtualizedGrid/headerCell';
 import type {ErrorFrame} from 'sentry/utils/replays/types';
-import useUrlParams from 'sentry/utils/useUrlParams';
+import useUrlParams from 'sentry/utils/url/useUrlParams';
 
 const SortStrategies: Record<string, (row: ErrorFrame) => any> = {
   id: row => row.data.eventId,
