+++ static/app/views/replays/detail/network/details/index.tsx
@@ -2,8 +2,8 @@ import {Fragment} from 'react';
 
 import DetailsSplitDivider from 'sentry/components/replays/virtualizedGrid/detailsSplitDivider';
 import type {SpanFrame} from 'sentry/utils/replays/types';
+import useUrlParams from 'sentry/utils/url/useUrlParams';
 import type {useResizableDrawer} from 'sentry/utils/useResizableDrawer';
-import useUrlParams from 'sentry/utils/useUrlParams';
 import NetworkDetailsContent from 'sentry/views/replays/detail/network/details/content';
 import type {TabKey} from 'sentry/views/replays/detail/network/details/tabs';
 import NetworkDetailsTabs from 'sentry/views/replays/detail/network/details/tabs';
