+++ static/app/views/replays/detail/network/details/tabs.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import {TabList, Tabs} from 'sentry/components/core/tabs';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import useUrlParams from 'sentry/utils/useUrlParams';
+import useUrlParams from 'sentry/utils/url/useUrlParams';
 
 const TABS = {
   details: t('Details'),
